write tests and fix bugs

+++ backend/src/main/java/ch/michu/tech/swissbudget/app/service/TagService.java
@@ -5,6 +5,7 @@ import ch.michu.tech.swissbudget.app.dto.tag.TagDto;
 import ch.michu.tech.swissbudget.app.dto.tag.UpdateTagDto;
 import ch.michu.tech.swissbudget.app.entity.CompleteTransactionEntity;
 import ch.michu.tech.swissbudget.app.exception.KeywordAlreadyExistsException;
+import ch.michu.tech.swissbudget.app.exception.TagAlreadyExistsException;
 import ch.michu.tech.swissbudget.app.provider.KeywordProvider;
 import ch.michu.tech.swissbudget.app.provider.TagProvider;
 import ch.michu.tech.swissbudget.app.provider.TransactionProvider;
@@ -43,11 +44,19 @@ public class TagService {
     }
 
     public void createTag(UpdateTagDto toCreate) {
+        toCreate.getKeywordsToAdd().forEach(this::validateNewKeyword);
         UUID userId = supportProvider.get().getUserIdOrThrow();
-        tagProvider.insertCompleteTag(userId, toCreate.getName(), toCreate.getColor(), toCreate.getIcon(), toCreate.getKeywordsToAdd());
+        UUID tagId = toCreate.getTagId() == null ? UUID.randomUUID() : toCreate.getTagId();
 
-        List<KeywordRecord> addedKeywords = keywordProvider.selectKeywordsByTagId(userId, toCreate.getTagId());
-        mapper.handleKeywordsAdded(userId, toCreate.getTagId(), addedKeywords);
+        if (tagProvider.fetchExists(userId, tagId, toCreate.getName())) {
+            throw new TagAlreadyExistsException(toCreate.getName());
+        }
+
+        tagProvider.insertCompleteTag(userId, tagId, toCreate.getName(), toCreate.getColor(), toCreate.getIcon(),
+            toCreate.getKeywordsToAdd());
+
+        List<KeywordRecord> addedKeywords = keywordProvider.selectKeywordsByTagId(userId, tagId);
+        mapper.handleKeywordsAdded(userId, tagId, addedKeywords);
     }
 
     // not allowed to update keywords, would be waaaay too much pain to make sure all transactions are correct -> user has to delete and add again
@@ -57,6 +66,10 @@ public class TagService {
         if (!tagProvider.fetchExists(userId, toUpdate.getTagId())) {
             throw new ResourceNotFoundException("tag", toUpdate.getTagId());
         }
+        if (tagProvider.fetchExists(userId, toUpdate.getTagId(), toUpdate.getName())) {
+            throw new TagAlreadyExistsException(toUpdate.getName());
+        }
+        toUpdate.getKeywordsToAdd().forEach(this::validateNewKeyword);
 
         tagProvider.updateTag(userId, toUpdate.getTagId(), toUpdate.getName(), toUpdate.getColor(), toUpdate.getIcon());
         keywordProvider.deleteKeywordsByIds(userId, toUpdate.getKeywordIdsToDelete());
@@ -66,8 +79,13 @@ public class TagService {
         mapper.handleKeywordsAdded(userId, toUpdate.getTagId(), addedKeywords);
     }
 
-    public void deleteTag(UUID tagId) { // FIXME can't just delete tag, need to also insert default tag into transactions with no tags
-        tagProvider.deleteById(supportProvider.get().getUserIdOrThrow(), tagId);
+    public void deleteTag(UUID tagId) {
+        UUID userId = supportProvider.get().getUserIdOrThrow();
+
+        UUID defaultTagId = tagProvider.selectDefaultTagId(userId);
+        transactionProvider.updateTransactionsByTagWithDefaultTag(tagId, defaultTagId);
+
+        tagProvider.deleteById(userId, tagId);
     }
 
     public void validateNewKeyword(String keyword) {
@@ -120,6 +138,9 @@ public class TagService {
         RequestSupport support = supportProvider.get();
         UUID userId = support.getUserIdOrThrow();
 
+        if (keyword != null) {
+            validateNewKeyword(keyword);
+        }
         if (!transactionProvider.fetchExists(userId, transactionId)) {
             throw new ResourceNotFoundException("transaction", transactionId);
         }
@@ -134,23 +155,11 @@ public class TagService {
         }
     }
 
-    public void changeTag(UUID transactionId, UUID tagId) {
-        UUID userId = supportProvider.get().getUserIdOrThrow();
-
-        if (!transactionProvider.fetchExists(userId, transactionId)) {
-            throw new ResourceNotFoundException("transaction", transactionId);
-        }
-        if (!tagProvider.fetchExists(userId, tagId)) {
-            throw new ResourceNotFoundException("tag", "" + tagId);
-        }
-
-        transactionProvider.updateTransactionWithTagAndRemoveNeedAttention(transactionId, tagId);
-    }
-
     private void addKeyword(RequestSupport support, UUID tagId, String keyword) {
         support.logInfo(this, "adding keyword to tag: %s->%s", tagId, keyword);
         UUID userId = support.getUserIdOrThrow();
-        UUID newKeywordId = keywordProvider.insertKeywordToTag(userId, tagId, keyword);
+        UUID newKeywordId = UUID.randomUUID();
+        keywordProvider.insertKeywordToTag(userId, newKeywordId, tagId, keyword);
         KeywordRecord newKeyword = keywordProvider.newRecord();
         newKeyword.setId(newKeywordId);
         newKeyword.setUserId(userId);
